backend/requirements.txt

fastapi
uvicorn
pandas
python-multipart
sqlalchemy
google-generativeai
python-dotenv
pydantic
psycopg2-binary
